+++ server/src/Providers/Builders/CompletionItemBuilder.ts
@@ -27,7 +27,7 @@ export default class CompletionItemBuilder extends Builder {
   }
 
   public static buildResolvedItem(item: CompletionItem, serverConfig: ServerConfiguration): CompletionItem {
-    if (serverConfig.completion.addParamsToFunctions && item.kind === CompletionItemKind.Function) {
+    if (serverConfig.completion.addParamsToFunctions && item.kind === CompletionItemKind.Function && !item.label.includes("(")) {
       const params = item.data as FunctionParamComplexToken[];
 
       return {
@@ -41,9 +41,9 @@ export default class CompletionItemBuilder extends Builder {
     return item;
   }
 
-  public static buildItem(token: ComplexToken): CompletionItem {
+  public static buildItem(token: ComplexToken, implicitConstants = false): CompletionItem {
     if (this.isConstantToken(token)) {
-      return this.buildConstantItem(token);
+      return token.isConst || implicitConstants ? this.buildConstantItem(token) : this.buildVariableItem({ ...token, tokenType: CompletionItemKind.Variable });
     } else if (this.isVariableToken(token)) {
       return this.buildVariableItem(token);
     } else if (this.isFunctionParameterToken(token)) {
@@ -80,7 +80,7 @@ export default class CompletionItemBuilder extends Builder {
   private static buildFunctionParamItem(token: FunctionParamComplexToken): CompletionItem {
     return {
       label: token.identifier,
-      kind: token.tokenType,
+      kind: CompletionItemKind.Variable,
       detail: `(param) ${token.identifier}: ${this.handleLanguageType(token.valueType)}`,
     };
   }