+++ server/src/Providers/SignatureHelpProvider.ts
@@ -1,7 +1,6 @@
-import { SignatureHelpParams } from "vscode-languageserver/node";
+import { CompletionItemKind, SignatureHelpParams } from "vscode-languageserver/node";
 
 import type { ServerManager } from "../ServerManager";
-import type { FunctionComplexToken } from "../Tokenizer/types";
 import { SignatureHelpBuilder } from "./Builders";
 import Provider from "./Provider";
 
@@ -21,17 +20,13 @@ export default class SignatureHelpProvider extends Provider {
 
       const context = this.getDocumentContext(uri, position);
       if (!context) return;
-      const { document, localScope, lines, rawTokenizedContent } = context;
+      const { lines, rawTokenizedContent } = context;
       const call = this.server.tokenizer.getCallContextFromRaw(lines, rawTokenizedContent, position);
       if (!call) return;
       const { identifier: rawContent, activeParameter } = call;
-      const functionComplexToken =
-        localScope.functionsComplexTokens.find((token) => token.identifier === rawContent) ||
-        document.getGlobalComplexTokens().find((token) => token.identifier === rawContent) ||
-        this.getStandardLibComplexTokens(uri).find((token) => token.identifier === rawContent);
-
-      if (functionComplexToken) {
-        return SignatureHelpBuilder.buildFunctionItem(functionComplexToken as FunctionComplexToken, activeParameter);
+      const functionComplexToken = this.resolveValue(context, rawContent)?.token;
+      if (functionComplexToken?.tokenType === CompletionItemKind.Function) {
+        return SignatureHelpBuilder.buildFunctionItem(functionComplexToken, activeParameter);
       }
     };
   }